# Library/Swagger/docs/swagger.yaml
basePath: /employee
definitions:
...
  Dto.TransferDataDTO:
    properties:
      createdBy:
        type: string
      createdTime:
        type: string
      deleted:
        type: boolean
      deletedBy:
        type: string
      deletedTime:
        type: string
      employeeCode:
        type: string
      employeeId:
        type: integer
      employeeName:
        type: string
      fingerPrintId:
        type: string
      id:
        type: integer
      machineId:
        type: integer
      machineName:
        type: string
      modifiedBy:
        type: string
      modifiedTime:
        type: string
      newCompanyLocationCode:
        type: string
      newCompanyName:
        type: string
      newDepartmentName:
        type: string
      newEmployeeCode:
        type: string
      newIdCompany:
        type: integer
      newIdDepartment:
        type: integer
      newIdLocation:
        type: integer
      newIdPosition:
        type: integer
      newIdSection:
        type: integer
      newLocationName:
        type: string
      newPositionName:
        type: string
      newSectionName:
        type: string
      oldCompanyLocationCode:
        type: string
      oldCompanyName:
        type: string
      oldDepartmentName:
        type: string
      oldEmployeeCode:
        type: string
      oldIdCompany:
        type: integer
      oldIdDepartment:
        type: integer
      oldIdPosition:
        type: integer
      oldIdSection:
        type: integer
      oldIdlocation:
        type: integer
      oldLocationName:
        type: string
      oldPositionName:
        type: string
      oldSectionName:
        type: string
      reason:
        type: string
      transferCode:
        type: string
      transferDate:
        type: string
    required:
    - employeeId
    - fingerPrintId
    - machineId
    - machineName
    - newCompanyLocationCode
    - newCompanyName
    - newDepartmentName
    - newEmployeeCode
    - newIdCompany
    - newIdDepartment
    - newIdLocation
    - newIdPosition
    - newIdSection
    - newLocationName
    - newPositionName
    - newSectionName
    - oldCompanyLocationCode
    - oldCompanyName
    - oldDepartmentName
    - oldEmployeeCode
    - oldIdCompany
    - oldIdDepartment
    - oldIdPosition
    - oldIdSection
    - oldIdlocation
    - oldLocationName
    - oldPositionName
    - oldSectionName
    - transferCode
    type: object
...
  Response.RespErrorStruct:
    properties:
      message:
        type: string
      result:
        additionalProperties:
          type: string
        example:
          key: value
          key2: value2
        type: object
      status:
        example: 400
        type: integer
      title:
        example: bad_request
        type: string
    type: object
  Response.RespResultStruct:
    properties:
      message:
        type: string
      result:
        additionalProperties:
          type: string
        example:
          key: value
          key2: value2
        type: object
      status:
        type: integer
      title:
        type: string
    type: object
info:
  contact: {}
  description: 'Example: Bearer abcdefghijklmnopqrstuvwxyz12345678901'
  title: Employee Service
  version: "1.0"
paths:
...
  /transfers:
    get:
      consumes:
      - application/json
      description: 'Sample Parameter: `?sortColumn=id&sortOrder=desc&pageSize=0&pageNumber=0&employeeCode=empl&transferCode=12`'
      parameters:
      - description: sortColumn
        in: query
        name: sortColumn
        type: string
      - description: sortOrder
        in: query
        name: sortOrder
        type: string
      - description: pageSize
        in: query
        name: pageSize
        type: integer
      - description: pageNumber
        in: query
        name: pageNumber
        type: integer
      - description: transferCode
        in: query
        name: transferCode
        type: string
      - description: employeeCode
        in: query
        name: employeeCode
        type: string
      - description: employeeName
        in: query
        name: employeeName
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/Response.RespResultStruct'
        "500":
          description: desc
          schema:
            $ref: '#/definitions/Response.RespErrorStruct'
      security:
      - Bearer: []
      summary: TransferData
      tags:
      - Transfer
    post:
      consumes:
      - application/json
      description: 'Sample Payload: `{"transferCode":"001","transferDate":"2022-04-01T07:48:48.243Z","employeeId":2,"employeeCode":"10.0002","machineId":1,"machineName":"nama
        mesin","fingerPrintId":"finger 1","oldEmployeeCode":"000.0002","oldCompanyName":"SSSS","oldLocationName":"Medan","oldDepartmentName":"Acconounting","oldSectionName":"Bagian
        1","oldPositionName":"Staff IT","oldCompanyLocationCode":"101010","newEmployeeCode":"10.0002","newCompanyName":"S4","newLocationName":"Medan","newDepartmentName":"MIS
        / IT","newSectionName":"Bagian 3","newPositionName":"Dev Ops","newCompanyLocationCode":"10101010","reason":"Reason"}`'
      parameters:
      - description: Example Payload
        in: body
        name: '"payload"'
        required: true
        schema:
          $ref: '#/definitions/Dto.TransferDataDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/Response.RespResultStruct'
        "500":
          description: desc
          schema:
            $ref: '#/definitions/Response.RespErrorStruct'
      security:
      - Bearer: []
      summary: TransferCreate
      tags:
      - Transfer
  /transfers/{id}:
    get:
      consumes:
      - application/json
      description: 'Sample Parameter: `3`'
      parameters:
      - description: Id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/Response.RespResultStruct'
        "500":
          description: desc
          schema:
            $ref: '#/definitions/Response.RespErrorStruct'
      security:
      - Bearer: []
      summary: TransferDetail
      tags:
      - Transfer
securityDefinitions:
  Bearer:
    in: Header
    name: Authorization
    type: apiKey
swagger: "2.0"